requirements.txt
pyspark==3.5.0
delta-spark==3.2.0
pandas>=2.0.0
pyarrow>=14.0.0
numpy>=1.24.0	
boto3>=1.34.0
requests>=2.31.0
python-dotenv>=1.0.0
apache-airflow-providers-apache-spark==4.9.0